// src/types/brewery.ts
export type Brewery = {
    id: string;
    name: string;
    brewery_type: string;
    street: string;
    address_2: string;
    address_3: string;
    city: string;
    state: string;
    county_province: string;
    postal_code: number;
    country: string;
    longitude: number;
    latitude: number;
    phone: string;
    website_url: string;
}